JOE-870: Document dual control-plane tax and extract runtime delta coalescer
Record dual memory+Postgres stores as accepted cost with facade shrink
guidance, and move the assistant.message delta coalescer out of app.ts
bootstrap into runtime-delta-coalescer.ts under modularity ratchets.
