+++ docs/architecture.md
@@ -554,3 +554,5 @@ Upstream promise: every `v*` tag on this repo corresponds to
 a tested SDK pair. Forks that track our tags inherit that
 guarantee; forks that live off `master` own their own bisect if
 a drift lands between tags.
+
+See also [Cloud dual control-plane tax](cloud-dual-control-plane.md).
